# backend/spec/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/workspaces/{workspaceId}": {
            "get": {
                "operationId": "WorkspaceScopedController_getWorkspace",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "timezone": {
                                            "type": "string"
                                        },
                                        "baseCurrency": {
                                            "type": "string"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "lastPaymentSourceId": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "baseCurrencyLocked": {
                                            "type": "boolean"
                                        },
                                        "role": {
                                            "type": "string",
                                            "enum": ["member", "admin", "owner"]
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "name",
                                        "type",
                                        "timezone",
                                        "baseCurrency",
                                        "createdAt",
                                        "updatedAt",
                                        "baseCurrencyLocked"
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["WorkspaceScoped"]
            },
            "put": {
                "operationId": "WorkspaceScopedController_updateWorkspace",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWorkspaceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "timezone": {
                                            "type": "string"
                                        },
                                        "baseCurrency": {
                                            "type": "string"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "lastPaymentSourceId": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "baseCurrencyLocked": {
                                            "type": "boolean"
                                        },
                                        "role": {
                                            "type": "string",
                                            "enum": ["member", "admin", "owner"]
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "name",
                                        "type",
                                        "timezone",
                                        "baseCurrency",
                                        "createdAt",
                                        "updatedAt",
                                        "baseCurrencyLocked"
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["WorkspaceScoped"]
            }
        },
...
    },
    "info": {
        "title": "Monqom API",
        "description": "The Monqom API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "UpdateWorkspaceDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 100
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "base_currency": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 3
                    }
                },
                "required": ["timezone"]
            },
...
        }
    }
}